refactor: 패키지 명 변경 exceoption -> exception
