+++ src/main/java/com/mju/lighthouseai/domain/board/service/impl/BoardServiceImpl.java
@@ -3,15 +3,12 @@ package com.mju.lighthouseai.domain.board.service.impl;
 import com.mju.lighthouseai.domain.board.dto.service.request.BoardCreateServiceRequestDto;
 import com.mju.lighthouseai.domain.board.dto.service.request.BoardUpdateServiceRequestDto;
 import com.mju.lighthouseai.domain.board.entity.Board;
-import com.mju.lighthouseai.domain.board.exceoption.BoardErrorCode;
-import com.mju.lighthouseai.domain.board.exceoption.NotFoundBoardException;
+import com.mju.lighthouseai.domain.board.exception.BoardErrorCode;
+import com.mju.lighthouseai.domain.board.exception.NotFoundBoardException;
 import com.mju.lighthouseai.domain.board.mapper.service.BoardEntityMapper;
 import com.mju.lighthouseai.domain.board.repository.BoardRepository;
 import com.mju.lighthouseai.domain.board.service.BoardService;
 import com.mju.lighthouseai.domain.board.dto.service.respone.BoardReadAllServiceResponseDto;
-import com.mju.lighthouseai.domain.cafe.dto.service.response.CafeReadAllServiceResponseDto;
-import com.mju.lighthouseai.domain.cafe.entity.Cafe;
-import com.mju.lighthouseai.domain.cafe.exception.CafeErrorCode;
 import com.mju.lighthouseai.domain.cafe.exception.NotFoundCafeException;
 import com.mju.lighthouseai.domain.user.entity.UserRole;
 import com.mju.lighthouseai.domain.user.exception.NotFoundUserException;
